requirements.txt
numpy
scipy
matplotlib
opencv-python
Pillow